+++ foxes/core/data.py
@@ -567,7 +567,7 @@ class Data(Dict[str, np.ndarray]):
         self.extra_data = deep_update(self.extra_data, shared.extra_data)
 
     @classmethod
-    def from_dataset(cls, ds, *args, callback=None, s_states=None, copy=True, **kwargs):
+    def from_dataset(cls, ds, *args, callback=None, s_states=None, copy=True, n_turbines=None, **kwargs):
         """
         Create Data object from a dataset
 
@@ -584,6 +584,8 @@ class Data(Dict[str, np.ndarray]):
             Slice object for states
         copy: bool
             Flag for copying data
+        n_turbines: int, optional
+            The number of turbines, if not found in the dataset
         kwargs: dict, optional
             Additional parameters for the constructor
 
@@ -595,6 +597,8 @@ class Data(Dict[str, np.ndarray]):
         """
         data = {}
         dims = {}
+
+        TODO NTURBINES
 
         for c, d in ds.coords.items():
             if c == FC.STATE:
@@ -792,7 +796,7 @@ class FData(Data):
         return data
 
     @classmethod
-    def from_dataset(cls, ds, *args, mdata=None, callback=None, **kwargs):
+    def from_dataset(cls, ds, *args, mdata=None, callback=None, n_turbines=None, **kwargs):
         """
         Create Data object from a dataset
 
@@ -807,6 +811,8 @@ class FData(Data):
         callback: Function, optional
             Function f(data, dims) that manipulates
             the data and dims dicts before construction
+        n_turbines: int, optional
+            The number of turbines, if not found in the dataset
         kwargs: dict, optional
             Additional parameters for the constructor
 
@@ -825,8 +831,13 @@ class FData(Data):
                     data[FC.STATE] = mdata[FC.STATE]
                     dims[FC.STATE] = mdata.dims[FC.STATE]
                 if FC.TURBINE not in data:
-                    data[FC.TURBINE] = mdata[FC.TURBINE]
-                    dims[FC.TURBINE] = mdata.dims[FC.TURBINE]
+                    if FC.TURBINE in mdata:
+                        data[FC.TURBINE] = mdata[FC.TURBINE]
+                        dims[FC.TURBINE] = mdata.dims[FC.TURBINE]
+                    else:
+                        assert n_turbines is not None, "n_turbines must be provided if not found in mdata"
+                        data[FC.TURBINE] = np.arange(n_turbines)
+                        dims[FC.TURBINE] = (FC.TURBINE,)
                 if callback is not None:
                     callback(data, dims)